Add port_type() method to server listen_socket and children
We're doing a significant refactor on the listen_socket and its
child classes.  There's a lot of duplicated code, and we can clean
up the flow a lot by moving some stuff around.

We generate a handful of log messages which have the socket type
in them, but since we are moving some of that functionality into
the listen_socket, we need to know from each derived type what
kind of socket they are.  port_type() returns a simple string with
that type in it.