Add do_logout() method to server listen_socket and children
The reaper worker static methods in both of the listen_socket's
child objects is almost exactly identical, with some small changes
to take care of the derived user types for each.  We can move that
specific functionality out into separate methods in each derived
class, so we can move the main reaper worker into the parent class.

This commit is those simple derived-specific functions, with a
pure-virtual in the parent class to ensure they are there.