Add tunable reaping parameters to the listen_socket constructor
We need to do some testing of the new reaper function, and instead
of having to wait around at least 75 seconds, we should be able to
shorten the test to a much shorter time.

It's a bit unfortunate that the derived classes need to have the
same default argument business as the base class, but it's not a
huge hardship.

We wouldn't expect to alter these values in practice, which is why
they're constructor-only inputs.