Move thread pool destruction after reaper destruction
It doesn't really matter which of the threads gets stopped first,
so order isn't important to execution.  However it's easier to test
the reaper thread if it gets done first, since it will throw the
exceptions we want to test, rather than some complicated mocking
to get pthread_cancel() or pthread_join() to ignore everything but
the reaper.  Thread pools get tested on their own, so we know they
work as expected.