Remove unneeded test for server listen_socket
We have tests which test create/delete, and start/stop.  The test
which tested a failure during the destructor, which was really a
failure during the stop method, didn't work correctly, and also
didn't test anything particularly meaningful.

We can revisit this test at a later time, as there will be one
untested line in the destructor as a result of the deletion.

The broken_listen_socket derived class was not used by any other
test, so we can get rid of it as well.