Resolve errors in server listen_socket reaper worker
We were using a variable to reference our base_user pointer, but
never actually initialized it, so we were getting instant
segfaults.

At the end of the user deletion, we were doing some complicated
stuff that we can make a lot simpler by splitting it into two
statements.  The last line, an erasure of the object that we're
operating on, invalidates the iterator we're using, so we need to
move our iterator to a place that (a) won't be invalid following
the erase, and (b) will allow the loop iteration operation,
incrementing the iterator, to continue correctly without missing
any elements.