Add some mocking in preparation for testing listen_socket reaper
The reaper sleeps a lot, so we need to mock out sleep.  It's not
nearly as crazy as it sounds, and will allow our test to run very
quickly.

Also, we'll need to construct more than one listen_socket, so we'll
copy in our function to make an addrinfo object that won't collide
with other sockets in the same file.

Finally, we'll need the config object, and we'll also need the
database pointer and stuff, so we'll include our mocked server
globals header.