Add beginnings of test for the server dgram socket
Since the datagram socket is a bit of a cornerstone to the entire
server program, we need to make sure it does what we think it does.
That means testing.

The dgram_user class, which adds a few members to the base_user
class, is where we start.

Re:  issue #163