Eliminate the base_user init() method
There were some articles I was reading around the time which I
initially wrote some of the network handling, which suggested that
initialization of objects should always be performed in an init()
method.  Cargo-cult programming is almost never a good thing, and
in this case, it led to unnecessary stuff, so we'll just move the
initialization into the constructor, because it just makes most
sense there.