Alter the packet_list to contain a user pointer
The packet_list structure is something that we pass around to keep
a socket, a user, and a packet all together, and since some of the
methods for sending packets now uses the whole user, rather than
just the control object, we'll swap them out.  The changes
cascade into a number of other places.

Eventually, we really should try to get rid of the packet_list
structure altogether (along with most of the rest of the defs.h
file).  It was created at a time when I wasn't sure what we would
be passing around, or how the individual data elements we needed
would be made available.  We're still not 100% sure, so for now it
will remain.