Move listen socket pointer from packet_list into base_user
We're working on making more sensible operation for the users, and
moving some of the functionality that resides with the sockets into
those users.  Since a given user will always be owned by a given
socket, we can just have a pointer to the parent socket.  This
simplifies a lot of things, not the least of which is the ability
to remove the socket pointer from the packet_list structure.

It's not obvious how we would remove the rest of the packet_list,
but this is at least one small step toward that.