Break packet handlers out of dgram listen loop
We're going to do a table-dispatch on the packets we get within
the datagram socket's listening loop.  We need to first break the
individual operations out into their own functions.  Since we can't
do function pointers to member functions, and testing plain static
functions is a bit of a hassle, we'll make them static class
methods.  That way, we can test them directly.