Use the map dispatcher in the server dgram_socket
Now that we have the map dispatcher, we can get rid of the complex
inline stuff that makes our loop hard to understand.

We can also alter the access_list structure a little bit; since we
have a Sockaddr object available by the time that structure comes
into play, we can use it directly.