Duplicate the Sockaddr during login in server dgram socket
After we do the dispatch in our server datagram socket, we delete
the Sockaddr object that we created.  The only time we need to
keep that object around is during the login, so instead of some
complicated logic, we'll just make a copy where we need to hang
onto it.