Add test for server dgram socket action request handler
The last of the static class packet handler functions is the action
request handler.  It requires an action pool to do the test, but
creating an actual action pool will be way too difficult.  We don't
actually need the full functionality of the action pool, but rather
just a basic thread pool.

Re:  issue #163