Add test framework for server dgram_socket workers
We're going to test the dgram_socket workers separately from the
rest of the methods, just because the mocking is going to be pretty
extreme.

The listener test can be driven entirely by recvfrom(2), and can
even be called directly.  No need for sleeping or busy-waiting.

The sender test can be set up fairly easily - load up the send
queue and start up the thread.

Re:  issue #163