Add mocking for dgram send loop test
The send loop for the server dgram_socket is a little more
complicated to mock, but not extremely so.  The function acts
progressively, so we can fail each call once, and then succeed the
rest of the calls.  hton_packet() bombs the first time, and then
works, as does sendto().

Re:  issue #163