Factor the user connection handling from stream listen loop
The last bit of refactoring we can do in the server stream socket
listening loop is the handling of data we may get from any of the
subservers.

This may be weird if a given subserver tries to write multiple
items between the time we handle one request and when we call
select again, so this may need some additional work.