Add test for server stream socket select_fd_test method
Our new refactoring split out some of the listening loop, both so
that we could more easily see and understand each piece, but also
because it's easier to test in this way.  This is the first of the
pieces that we split out, and the test is fairly simple.

Re:  issue #162