Add the fds map into the server stream socket
Our process is heavily threaded, so each thread can handle a lot
of computation.  Processes on modern Unix hosts can handle gobs of
sockets, and with our threading model, we should be able to handle
a whole lot of stream users in our main process.  Our subserver
model for maximizing the number of stream users seems a little
antiquated, so we should just get rid of it, both for the sake of
simplicity of code, and for the ease of testing.

First step is to start moving file descriptors out of the subserver
and into the main stream socket.  As with the datagram socket, we
can put them into a map which will give us quick lookups.  The
important part of stream user handling is the file descriptor,
rather than the sockaddr, so our new data structure will be a map
of file descriptors, which relate to stream user objects.