Convert port_type method into plain virtual in server listen_socket
In issue #168, we found a flapping test, which seems to have been
caused by some incorrect understanding of vtables and virtual
destructors.

We were calling out to stop() in the listen_socket destructor,
which in turn may have called out to port_type().  At that point
in any of the derived classes' destructions, those virtuals would
not be available, so the only vtable entries that were left were
those of the listen_socket itself.  Since port_type() was a pure-
virtual, blammo.

It's interesting that it was a flapping error, rather than a
consistent one.

We've updated the unit test as well, to remove the override from
our test derived class, and to test the port_type() method
directly.