Add create/delete test for server control object
We don't have a set of tests for the actual control object.  We
test it indirectly through the rest of the test suite, but that's
not the same as exercising it directly.

We're about to make some changes to it, so we need a good set of
tests to go along with the changes.