Add equality operator for server control
The other operator we're likely to need is the equality comparison
operator.  This does a pretty shallow comparison, and is mostly
just for sorting.  Based on the way the STL seems to work, it's
unlikely that we even need this, but for completeness' sake, we'll
have it.

Also added a unit test for this operator.