Derive server base_user from the control object
All the user types have a pointer to the control object, which is
silly because the user is mostly worthless without the control.
Instead, we'll just make the user objects into control objects by
deriving from the control.

There's going to have to be some refactoring that goes on to make
everything work again, but the merging of the two objects is done.