Add descriptor/sockaddr maps to derived socket classes
We're restructuring the way our sockets operate.  We're going to
have only the base_user, and the dgram_user/stream_user types will
go away.  In order to map a base_user to a descriptor/sockaddr, we
need objects that go both ways - userid --> thing, and thing -->
userid.

This commit adds the userid --> thing maps, and adds maintenance of
the data in them.

Re:  issue #173