Add overridable methods for user connection/disconnection
In order to get rid of some of the pure-virtual methods which have
caused trouble in our test suite, we'll have methods which do some
of the listen_socket-specific parts of user login and logout,
notably adding and removing the user from the listen_socket user
maps, which can be overridden in the derived classes.

Each derived class will want to call the parent's connect_user/
disconnect_user method at the end of its own, to make sure the data
structures maintain integrity.

Re:  issue #173