Rework login process in server listen_socket
There are several things at work here.  First, we want to get rid
of the dgram- and stream-specific user classes, so the login
process will operate on only base_user objects.

Second, we want to make sure that the default slave of any user's
object is set to their character object - we've got some control
behaviour that we never really used, that will be valuable once
people start trying to take control of things.

Next, we add the access level that the user has on this server
into the user object itself.  It may be useful for, say, action
routines to know what sort of access the user has.  Also it allows
us to pass the access level around without having more parameters.

Finally, we begin using the connect_user() method that we added
in the previous commit.

Re:  issue #173