Repair wait conditions in listen_socket access pool test
We were getting segfaults when trying to look for a given user in
the access pool test.  There should never be any users in the user
map, so that wasn't very smart.  Instead we'll just watch the work
queue for the access pool, and when it empties out, we've exercised
the stuff we need to exercise.

Re:  issue #173