Add generic packet_handler typedef in server listen_socket
Most of the packet handlers which currently exist within the dgram
socket can be moved into the listen socket, as they have nothing
whatever to do with the sockaddrs or file descriptors of the
derived classes.  We'll start with just the typedef for now.

The login will have to remain in the individual classes.