Do packet handling within the server stream_socket
We had not previously done any packet handling in the stream
socket, so we copied a similar process from the dgram socket.

We added a lookup map, which is used by the new handle_packet
method, and a stream-specific handle_login method.  We get the rest
of our packet handler functions from the listen_socket.