Remove unneeded delete call from server dgram_socket test
As we were copying the connect_user test from the server dgram
socket to the stream socket test, we noticed that we deleted the
database global, which we never even allocated in this test case.
Fortunately it never caused any trouble, but we'll get rid of it
all the same.