Add user disconnection test for server stream socket
Ideally, we would have a legitimate open file descriptor, and a
valid value for max_fd, but since we ignore the return value of
close(), it's not very important.

At some point in the future, we should at least log the fact that
close() has failed, so we can monitor descriptor leakage.

Re:  issue #162