Allow testing of private members of server stream_socket
We need to be able to do operations on the private members of the
stream_socket class in order to get coverage on all lines.  The
using declaration can change the access type of a member, so we'll
add a test stream socket with everything public.  Private members
may not be so changed, so they need to become protected.

Re:  issue #162