Clean up system calls in new-user method for stream_socket
There were some calls out to setsockopt() and ioctl() in the
stream_socket's accept_new_connection() method which were doing
their arguments incorrectly.  Most of the arguments to those
functions expect pointers to ints, not pointers to bools (which we
have no idea how they're represented internally).  So we add a
couple new local variables within the FD_ISSET() block, and do the
assignments correctly.

We also add a unit test for this method.  We have assertion
coverage on almost everything, but were unable to mock out the
ioctl() call successfully.  We don't check for error returns on
any of those calls, so it's not that critical.

Re:  issue #162