Improve read error handling in server stream_socket user handler
When we try to read a packet, and the return is either 0 or -1,
we'll consider that user too broken to continue.  Instead of doing
the cleanup there, we'll let the reaper do the work.

Also added a unit test to make sure this does what we think it
should do.

Re:  issue #162