Remove unnecessary extra test for server dgram_socket
The worker test was based on the fact that we previously would
cast our user into the appropriate type for the socket, but since
we no longer have specific user types for each socket, it is
completely unnecessary.

As such, we can remove the second test user, and just use the one.