Repair segfault in server stream_socket send loop
In previous incarnations, the stream_socket send loop would use a
pointer variable to ensure that it had the right kind of user.  Now
we just need to grab the file descriptor, and don't need to assign
the user to anything.

Also add a unit test for the send loop.

Re:  issue #162