Improve server listen_socket reaper loop
We were using a for loop, and doing adjustments of the loop index
in both directions because of the fact that we were deleting things
and invalidating the index.  Let's do the loop in a more straight-
forward way, and only move the index when we want to move it, in
the direction we want to move it.