Eliminate race between destructor and stream send worker test
We were seeing some segfaults on linux hosts, which appeared to be
a race between the send loop and the destructor.  We can ensure
that the send loop is done doing what we think it's doing by
waiting until our test index is one higher (which it will be by the
time we're done testing).