Add test for server stream_socket listen worker
All the pieces of the stream_socket listen worker are tested and
working, but the listen worker itself hasn't been tested.  Forcing
a specific value for the selected fd_set is the key here; not
clearing out the set made our functions assume that the listening
socket also got some data.

Re:  issue #162