Repair server basesock unix deletion test on Linux
We were running into a segfault on Linux hosts when trying to
capture a log message which occurs when unlinking of a unix domain
socket fails in the basesock destructor.  We temporarily reassign
the rdbuf of std::clog to a stringstream, so we can capture that
log message, and assign it back after the assertion is done.  The
segfault came in the destructor of the stringstream object.

Instead of spending time debugging the system STL for what we might
be doing wrong, we're going to allocate the stringstream on the
heap, and leak it out of the test, so the destructor likely never
happens.  We just get freed at the end of the program, as the
entire memory arena is freed.  Lame, but the problem isn't related
to the thing we're testing, and it gets around the failure.

Re:  issue #171