Add create/delete test for server octree
One of the first issues we ever added to our list, we need to add
a full test suite for the octrees that the server uses to keep
track of all the objects within its space.

First simple test is creation and deletion, with a couple simple
assertions to see what we created.

Re:  issue #7