Add some comments about our strategy with embedding python
Embedding python is not going to be a straight shot; I/O is handled
via regular stdin/stdout/stderr descriptors, and there's no
standard way to get output from some function that we might call.
We found somebody who had come up with a method to lasso the stdout
object and put it into a string, so we'll head in that direction.