Clean up composite mouse motion callback driver
We use a bit cleaner initialization and modification of our
variables within the mouse motion method.  We also clean up the
logic flow.  Since the composite is also an active object, it can
also have callbacks, and they should be called when the cursor
is not pointed at any children.