Shift the contained UI library to use cuddly-gl as a submodule
Since we have split the UI library into its own project and code
base, which is now more featureful than our existing library, we'll
strip out the old and replace it with the new.

There is work in reintegrating the library - directory names have
changed, etc.  This will come in subsequent commits.