Add second pure virtual method to get DB auth via charname
When performing a login, we're trying to use the server game object
ID in place of a character ID, which isn't working properly.  We'll
use the character name, which we do have at the time we're trying
to log the user in, in place of the character ID, which we can't
currently retrieve by any existing method.

We will add the database-specific methods in upcoming commits.