Remove unneeded forward declarations in server defs file
The type definition file which we so desperately want to get rid
of has several forward declarations of classes at the top, so we
can define pointers to the types in some of our defined types.  A
couple of those forward declarations are no longer needed, so we
can get rid of them.

One more step toward getting rid of this file once and for all.