Call zone method to send local objects on user login
When a user logs in, they need to have all the objects that are
nearby their current position, and the way to do that is call the
appropriate method in the zone.

The previous method did the full connection of the user to its
object, but since that's now handled within the listen socket, the
zone can focus on what it knows about:  where objects are.  We'll
change the name to reflect what the method now does.

This small change cascaded down through a few tests.

We should have a fully-working login process once again.