Send user's objectid back with the login acknowledgement
The client user never knows what their avatar's object ID is, since
we have no means to send it.  Now that we have more general-purpose
fields in the ack packet, we can provide that info to the client.

We'll save the object ID in the comm object, because the object ID
may be different for each server we're on.  We'll send that object
ID back with each of the action requests, since our avatar will
always be the source object.

Re:  issue #178