Accept no object case in user connection
We were getting some segfaults in our test suite during user
connection, and it ended up being that our test users didn't have
a slave game object.  We do want to support that configuration,
that a user can connect and look around, but not be associated with
any given avatar.