Ensure moves have a valid starting time
When we do a move, we scale the movement by the last time we
updated the object's position.  If we have an object that's sitting
still for 9 hours, and have it start moving, the movement will be
scaled by that 9-hour interval, rather than from the start of the
move, which will NOT give us the expected result.

We'll set the last-update time to the start of the move, so we get
correct scaling values once we get into the motion pool.