Add self-object pointer to client
There are a lot of instances in which we may need to do things
with the object that represents our user; we'll keep a global
pointer to that object.

The move callback can now get the orientation of our object (which
will always be nothing at the moment), and move in that direction.

Re:  issue #178