Import libtap++
We don't build or use it yet, but it's here.  Our CI platform
will need to have the libboost-dev package installed in order
to build it, so we'll go ahead and install it now.

Also, we missed removing the Xerces library when we removed the
dependencies on Xerces, so we'll do it while we're in the file.