Convert client comm socket exception test to use libtap++
We're converting from gtest/gmock to use libtap++.  We'll go test
by test, so we don't overwhelm things.  The client comm exception
test is pretty small, so is a good first conversion try.

We do add a bunch of gtest management stuff into the main function
to keep the unconverted tests working before we convert them.