Convert client comm start failure test to use libtap++
This is the last of the client comm exception tests to convert, so
we also remove the gtest dependencies from the file and build
configuration at this time.

Until all our test report in TAP format, automake's tap test runner
isn't appropriate.  For now, we'll just rely on the return value
from the exit_status() function, and if we need
assertion-by-assertion pass/fail, we can always look at the log
file for the test in question.