Convert server db success test to use libtap++
There is no assertion in this function, which may violate some sort
of testing maxim, but it will generate a bad thing if it doesn't
succeed, and should fail out the test.

This is the last of the db tests to be converted, so we can get rid
of the gtest stuff in the test and the build config for the test.