Convert client comm ping-receipt test to use libtap++
Using a regular derived class which simply makes protected methods
public makes the packet-receipt tests a whole lot easier to do, and
also allows us to get rid of some of the sleeping.

This is the first test in this file to convert, so this also
includes the makefile changes and scaffolding to make libtap++
work as expected.