Convert client comm position update receive test to use libtap++
Fortunately this function doesn't do a whole lot, so this test can
be pretty simple.  We do update the move_object function to check
what we're passing in, but we don't get the appropriate strings
in the build log.  We get the results, so we'll not worry about it
too much for now.