Convert client comm receive ntoh failure test to use libtap++
We've got a handful of tests which exercise exceptional conditions
in the receive handler, and they're all very slow.  We'll take a
different strategy:  they all log something, so instead of doing a
blind sleep, we'll wait until there is something in the log, and
then end the test.  Tests should end up running on the order of
milliseconds, instead of a second each.

This is for a failure in the ntoh_packet call, at the end of the
receive loop.