Move clog rdbuf changes into client comm test global scope
We replace the std::clog rdbuf in almost every single test, and
on some platforms, replacing the original rdbuf can cause
segfaults, so let's only do it once, in the global scope.  We set
it up at the start of the tests and clear it out when it needs
clearing.  We don't need to worry about resetting it, because
once we're done with it, the test is over.