Convert client comm hton failure test to use libtap++
Instead of sleeping while the threads do their work, we'll busy-
wait until the queue is empty, and continue on from there.

This is the last of the tests in this file to be converted, so
this change also removes the gtest/gmock includes, the gtest
artifacts in main, and in the build configuration.