Add fake DB for mocking purposes
Since we're removing the need for gmock, we'll need to make fake
objects that do nothing in place of the gmocked classes.  This is
a fake database object.  For now, it's only meant to mock out a
single method, but since almost all of the methods are pure-virtual
we have to implement something for them.

As we go along, we'll mock the methods that we need at the time
when we need them, to keep the diffs (and conversions) simpler.