Convert server listen_socket check_access test to use libtap++
This is a more complex test, of the full access-checking method.
We call out to the database quite a bit more, so we need to add
a bunch of mocks into the fake database object.  We also use a
fake zone object, so we need to make a new one that isn't handled
by gmock.