Convert server base_user ack test to use libtap++
This is another of the tests that was added later, in the wrong
place in the file, so we move it along with the conversion.

This is the last test in this file to be converted, so we remove
the rest of the gtest artifacts.  However, since we use the gmocked
database header, which still needs the include paths, we can't yet
get rid of all of the gtest build configs for the listensock test.