Remove unneeded gtest items
In the server dgram test, we reference a mocked object that isn't
even used, so we can get rid of it and the other gmock-related
stuff from the file.  In the server zone test, we missed removing
the line which initializes gtest, which we totally do not need
any more.