Add protected default constructor for server library
We've been seeing segfaults during construction of our fake
libraries, due to inability to mock out the open function
adequately.  We'll add a protected constructor which does nothing,
in order that we can call it in our fake derived classes, instead
of the constructor that expects to open a real library.