Remove last elements of gtest from build configs
We discovered the use of build-aux from the cuddly-gl project, and
it segregates the build scripts from the rest of the codebase, so
we'll do it here.  That moves the install-sh file into that
directory.

We add the tap-driver.sh script, which isn't normally installed as
part of the autogen process.  That also lives in build-aux.

The test makefiles now reference the tap-driver.sh script for the
test log driver.

There were a couple of small references to gtest and gmock which
we neglected to remove previously.