Add test for bad packet type
This is the first of the protocol tests, so we add all the extra
stuff:  ignore lines, build configs, the base test.

While looking where we should start testing, we did notice a
possible segfault, in which a too-large packet type would roll off
the end of our array.  We added some guards, and decided they would
be a fine place to start testing.