Add test for proto ping packet
The ping packet is just a basic packet with a specific type.  We
were not checking the size of our packet, which could have led to
some inadvertent segfaults, so we've added a guard clause.
Additionally, the basic packet no longer contained only byte-
sized values, so we did need to add some hton/ntoh calls.