Add configure option for GLM
We use the GLM includes for all our vector math.  Instead of
forcing the configure CXXFLAGS args to contain the right include
path for GLM, we'll add a configure option, --with-glm-includes.
This also helps with building cuddly-gl, as it expects the same
option, and will fail to configure if it needs it and doesn't get
it.