Handle the python GIL in our language module
The Global Interpreter Lock can be a bit of a hassle, and if we
handle it wrong, things will at best freeze, but more likely will
just error out.

PyEval_AcquireLock and PyEval_ReleaseLock have been deprecated,
and don't seem to do anything on the python I'm using.  They've
been replaced by the PyGILState_* functions, which present their
own set of challenges.

The initialization functions do create the GIL, but they leave it
locked on completion.  It's not clear how we should drop it,
without having gotten a PyGILState_STATE object previously.

Now that we're releasing the locks (mostly) reliably, we found that
if we try to call PyFinalize without the lock, we just fail.  That
is nowhere to be found in the python docu.  In fact, a lot of the
docu for this work is limited at best, and examples vary widely.