Add unit test for running a python script
Similarly with the other language modules, we'll make a function
called 'power', which computes a base raised to a power, and call
it via our object.

We currently have no mechanism to retrieve results from our python
object, so we'll have to be content with an empty string.

In future commits, we'll add the stdin/stdout/stderr override code,
as it's pretty complicated.  Just want to make sure the base stuff
is working before we wade into that ocean.