Provide a return value from the python execute method
Our embedded python interpreter had no way to return a value.
Unfortunately there's no direct way to return things from any of
the PyRun functions, so we have to set a particular variable in
the evaluated code, 'retval', and dig it out of the global
dictionary after we return.

At least it works.