Add implementation for embedded lua execute method
This is a bit more complex than some of the other embedded
languages, because we actually have types we need to contend with.
Fortunately, things get returned on the stack, so we can just nab
the top of the stack, do our checking, convert it to a string, and
return it.

Re:  issue #110