Add digest function for DH key exchanges
Standard use of the shared-secret-generation functions is that we
make a shared secret, and then feed it through some sort of digest
function to spread out distribution of shared keys across the key
space.  To that end, we'll make a function that takes a key of
some sort and takes the SHA256 digest of it.