Add string-to-key converter
We need to go the other way and convert a string back into a key
structure.  Now that we can reliably convert a key into a string,
we can take that output and convert it back.

In future commits, we should ensure that the key structure we get
back from the converter is the same as the structure we initially
sent into the key-to-string converter.

Re:  issue #31