Write private keys instead of public keys
Since EC keys can derive the public from the private, we don't
need to deal with public keys in general.  The EC functions all
seem to operate on a single key, so we'll treat that as a private
key.

The difference between the PUBKEY and PrivateKey functions is not
large, since we're not currently dealing with passphrases.