Test the server action_pool destructor symbol failure
When the server action_pool is destroyed, we try to get a symbol
out of our action library; if that symbol is not available, the
destructor should eat the exception that results.  We didn't have
test coverage of it before, but since we're in here, we might as
well add it.