Add default constructor to server basesock
When creating a fake basesock object during testing, we're running
into some occasional segfaults, which seem to be a result of some
of the complicated mocking involving the listen_socket.  One of the
complicated things that the listen_socket requires is to create a
addrinfo struct for the basesock.

In order that we don't necessarily have to make a fake addrinfo
structure, we'll have a default constructor which doesn't actually
set anything up.

The start method depends categorically on there being a socket, and
throws an exception if that's not the case, so we'll turn that into
a soft-failure, and simply have the start method not do anything in
that case.  It's unlikely that we'd get into that case during a
test, but we'll handle it just in case.