Use the default constructor for the fake stream test socket
In our stream socket unit test, there are a lot of conditions in
which we don't need to worry about having a valid socket underneath
the object.  We can switch to using our fake stream socket, which
skips that initialization, and eliminate a lot of the getaddrinfo
calls that we do.