Move freeaddrinfo calls closer to corresponding getaddrinfo calls
When we must construct an addrinfo structure, we should free up
the memory as soon as we can.  For a couple of our tests which do
require the underlying socket, we'll free up the addrinfo struct
right after we make the stream socket object - once the socket is
created, the addrinfo is unneeded.