Add some delay in server stream socket start/stop test
We've seen some hangs in the start/stop test for the server stream
socket.  The startup seems to complete; we get all of the expected
log entries, but the test stops completely at that point.  While
working to determine where the freeze was happening, we added an
output to cerr after the startup was complete, and the freezes
stopped immediately.  It appears that we weren't giving all the
threads a chance to start up fully before we tried to kill them.

This feels like a really hacky fix, but it is a very simple change,
and it does seem to repair the problem.

Re:  issue #190