Update cuddly with new set behaviour
The cuddly-gl project has made a significant set of changes in that
all the set methods now take direct values, instead of void
pointers.  Additionally, the set_va methods are gone, in favor of
a varargs templating arrangement that is much more in the C++
idiom.

Since our types are now more closely checked, we need to make sure
that we're calling a label; widgets don't have fonts or strings.
So we'll ensure that all of the log entries contain a label widget
instead of a generic widget.