Add a default log entry on log display construction
Having a single entry (that we can mostly ignore) in the log
display list simplifies a lot of the logic surrounding "what have
we displayed".  We can eliminate most of the complicated testing
against the end of the list, and can instead focus on what we have
actually not displayed.