Add printing out of failed ciphertexts in encrypt test
During development, we saw failures in some of our CI builds
regarding the output of the encryption funtion.  The key length was
wrong, but also we had no idea what the encryption was even
generating.  To that end, we'll add output of the ciphertext we
generated, and also what we expect it to be, to aid diagnosis of
whatever errors are happening.