Allow pkey_to_file to take no passphrase
If we don't care about having a passphrase on a private key file,
we should be able to feed in NULL, and have OpenSSL not worry about
encrypting the file.  It turns out that if we pass in NULL for the
cipher, the key file won't be encrypted.  Plus, we were running
into some segfaults when trying to take the strlen() of NULL.

To get some test coverage, we'll use the existing file_to_pkey
test.  That test also had a bunch of incorrect naming, so we'll fix
that while we're tweaking the test.