Generate the key file path in r9_keygen
When creating a key, we want to write it into the correct place,
depending on whether we're making one for a client or a server.

The client keeps it in the .r9 directory in the home directory, so
we'll grab the HOME environment.  Servers will keep them in their
/etc-ish directory, which we can set with various configure
options.