Add passphrases to r9_keygen
By default, we'll assume that clients do want an encrypted key file
and servers do not.  This would allow an unattended server startup.
Clients will input their password to decrypt the key file.

Since we haven't dealt with the termios stuff before, we need to
add checking for the header file into the configure script.