Add regeneration of po files to make all
We always forget to regenerate the localization po files when we
add new strings, so we'll make it (almost) always happen on a 'make
all'.

We still need to add filenames to the POFILES.in file manually, but
at least the po files themselves should always be in sync with
what's actually in the source files.