Add proto function to extract public key to DER
The ASCII-armored PKCS#8 format for public keys is readable by
humans, but it's more data than we really need to send across the
wire.  DER format is a lot more compact, and binary, so we should
have some functions to convert to and from it.

This function takes an EVP_PKEY and converts to a DER string.