Add public key to login request packet
We'll be moving from sending passwords to sending public keys
instead.  For now, we'll keep the password and do a parallel
implementation with the pubkeys.

Also, we found that the DER format, which is the most compact
representation of the key type we're using, takes up 170 bytes.
Our previous server key packet used 128, which is too small, so
we'll increase the size of that packet to 256 bytes as well.